requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# Vision Models
timm>=0.9.0
open-clip-torch>=2.23.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Dataset & Data Loading
pycocotools>=2.0.7
datasets>=2.14.0
albumentations>=1.3.1

# Training & Optimization
wandb>=0.15.0
tensorboard>=2.14.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Evaluation Metrics
nltk>=3.8.0
rouge-score>=0.1.2

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
